+++ include/ignition/sensors/CameraSensor.hh
@@ -125,6 +125,30 @@ namespace ignition
       /// \return Camera in Ignition Rendering.
       public: rendering::CameraPtr RenderingCamera() const;
 
+      /// \brief Topic where camera info is published.
+      /// \return Camera info topic.
+      public: std::string InfoTopic() const;
+
+      /// \brief Advertise camera info topic.
+      /// \return True if successful.
+      protected: bool AdvertiseInfo();
+
+      /// \brief Advertise camera info topic.
+      /// This version takes a string that allows one to override the
+      /// camera_info topic.
+      /// \param[in] _topic The topic on which camera info is to be published.
+      /// \return True if successful.
+      protected: bool AdvertiseInfo(const std::string &_topic);
+
+      /// \brief Populate camera info message.
+      /// \param[in] _cameraSdf Pointer to SDF object containing camera
+      /// information.
+      protected: void PopulateInfo(const sdf::Camera *_cameraSdf);
+
+      /// \brief Publish camera info message.
+      /// \param[in] _now The current time
+      protected: void PublishInfo(const ignition::common::Time &_now);
+
       /// \brief Create a camera in a scene
       /// \return True on success.
       private: bool CreateCamera();