+++ Changelog.md
@@ -2,12 +2,30 @@
 
 ### Ignition Sensors 2.X.X (201X-XX-XX)
 
-### Ignition Sensors 2.0.0 (2019-05-XX)
+### Ignition Sensors 2.2.0 (2019-06-27)
+
+1. Update the GPU Lidar to use the sensor's name as the `frame_id`.
+    * [Pull request 74](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/74)
+
+1. Fix camera_info topic to be on the same level as image and depth_image for RGBD Camera.
+    * [Pull request 73](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/73)
+
+### Ignition Sensors 2.1.0 (2019-06-18)
+
+1. Adds an RGBD camera sensor that combines a CameraSensor and DepthCameraSensor, and also
+   outputs a pointcloud.
+    * [Pull request 70](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/70)
+
+1. Create and publish on `camera_info` topics for the Camera and DepthCamera
+   sensors.
+    * [Pull request 67](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/67)
+
+### Ignition Sensors 2.0.0 (2019-05-21)
 
 1. Zero update rate, virtual SetParent and fix gpu_lidar
     * [Pull request 66](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/66)
 
-1. Add frame_id to sensor messages
+1. Add `frame_id` to sensor messages
     * [Pull request 63](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-requests/63)
 
 1. Restore `pixel_format` in message and add deprecation comment.